And move it to the index page
